requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
idna==3.4
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14